Modify binding of `this` and `super`
As states and transitions are inferred as Java classes, by default keywords
`this` and `super` are binded to the inferred Java class of the given state
or transition, and - if exists - its superclass respectively. This commit
modifies this binding so that the abovementioned keywords refer to the
enclosing XtxtUML class and - if exists - its superclass respectively.