Bind `super` only to XtxtUML superclasses
A small fix in the type resolver. If the enclosing Java class of a given
`super` expression doesn't have an actual Java superclass, Xbase starts to
look for an enclosing Java class which has one, and if such exists, binds
`super` to its superclass. This commit modifies the default behaviour such
that only enclosing XtxtUML classes are considered in the `super` lookup.