Add signal access ('sigdata') language element
Due to the modifications applied by 66e82d2, the api representation of the
signal access element (the 'getSignal' method) has become inaccessible in
XtxtUML. As its introduction as a language element is needed nevertheless,
this commit fixes the aforementioned issue by doing so.

In addition to the sole implementation of the 'sigdata' element,
validation regarding its usage has been added as well.