Improve type computation of `sigdata`
1. Prevent infinite loops
-------------------------
In case of one or more choice pseudostates being connected with
transitions such that a directed circle is formed, the previous
implementation would result in an infinite loop during the type
computation of any `sigdata` expression located inside any transition
starting from any of the abovementoined choice pseudostates. This
version fixes this issue.

2. Different type computation based on whether the given `sigdata`
   expression is located inside an entry or an exit activity
------------------------------------------------------------------
In the former case triggers of incoming, whereas in the latter case
triggers of outgoing transitions should be considered during the type
computation of the given `sigdata` expression.