Fix type usage collection in XtxtUML
Type usage collection in Xbase is mainly used to produce "unused import"
warning messages. However, if a type is referenced only in grammar-level
cross-references inside custom expressions, its usage is not considered.
This commit fixes this issue by extending the type usage collection to the
related XtxtUML expressions as well.