Dotted qualified names in "unused import" warnings
By default, generated Java equivalents of nested XtxtUML types (for
example association ends) are distinguished from their language-level
counterparts by using the `$` as separator instead of `.` between
enclosing and nested classes in qualified names. This notation is used in
issued "unused import" warnings as well. As it can be confusing for users,
this commit changes these warnings to use `.` uniformly in qualified
names.