Use FQ names as unique identifiers in XtxtUML
Several aspects of the previous implementation have been based on the
assumption that an XtxtUML EClass (e.g. a TUClass) instance can be used to
uniquely identify the related source element (e.g. an XtxtUML class
definition). Apparently, this is not the case.

In some situations -- mainly when a reference to a source element is used
outside of the defining file --, multiple EClass instances are created
and, as a result, these instances cannot be directly used to compare the
underlying definitions.

As a workaround, fully qualified names can be used as unique ids with the
notable exceptions of duplicate checks. Fortunately, it seems that direct
EClass instance comparison works fine in these cases.