Improve null-safety in XtxtUML
Xtend provides an operator for null-safe feature calls (`?.`). For
example, `foo?.bar` is an equivalent of `if (foo != null) foo.bar else
null`.

Now, consider the following example: `foo?.bar.baz`. You might think (as I
have) that this is translated to `if (foo != null) foo.bar.baz else null`.
Unfortunately, this is not the case. The correct translation: `(if foo !=
null) foo.bar else null).baz`. As it can be seen, if `foo` is null, the
second, not null-safed' feature call will result in a NullPointerException.

Therefore, if a null-safe feature call is present in a chain of feature
calls, all subsequent feature calls should be made null-safe.