Fixes and improvements in the content assist
1) By default, the scope provider is used to create content assist
proposals for cross-reference features. The acquired scope consists of
IEObjectDescriptions. In order to provide only relevant proposals, we
filter the given scope. In some cases, we have to investigate the
container of the referenced element (e.g. to decide whether an association
end belongs to a composition), but EObject descriptions don't always hold
such information. The method `getEContainerDescription`, implemented in
XtxtUMLUtils.xtend, provides a workaround for this issue.

2) In case of class property access expressions, properties of the source
object are now considered instead of properties of the expression's
enclosing class.

3) IEObjectDescription.getEObjectOrProxy.getFullyQualifiedName is an
unnecessary indirection, IEObjectDescription.getQualifiedName is now used
instead.