Fix JVM inference of enclosed members
Using the `accept` method of the JvmDeclaredTypeAcceptor during the
first phase of the JVM inference process adds the inferred JVM type to
the direct contents of the underlying EResource.

Although this aspect can be used to early-infer referable skeletons of
source elements -- which will be lazily initialized during the second
phase --, keeping accepted, but to-be-nested JVM types (e.g. states)
amongst direct contents of the aforementioned EResource apparently
causes anomalies such as issue 173.

The JVM inference process is started in the JvmModelAssociator's
`installDerivedState` method, which has a hardcoded reference to the
default JvmDeclaredTypeAcceptor. As it is quite long and complex, due
to maintainability reasons I have decided to implement an extension
method inside our model inferrer instead of defining a custom model
associator with an overriden, almost identical `installDerivedState`
method.