Use EContainers directly if necessary
Apparently, EObject descriptions tend to provide only a skeleton of the
described EObject -- which is accessible through the description's
`getEObjectOrProxy` method --, e.g. it's common that the fields of the
accessed EObject skeleton are `null`.

As a result, siblings of an association end description are not always
obtainable through its container description (see e3a4a6e for details
about why is it necessary to investigate container descriptions instead
of direct EContainers), therefore direct EContainers should be
considered as well in this case.