Provide launch shortcuts for XtxtUML executions
The necessary extension points were mainly copied from the plugin.xml
of the Xtend IDE project. However, as Eclipse effectively uses only
the first adapter factory registered for a given adapter type, the
JavaElementDelegateAdapterFactory injected by Xtend shadowed the one
injected by XtxtUML.

To overcome this issue, a custom adapter is defined, along with an
appropriate adapter factory. Due to the customized adapter, a custom
launcher is implemented as well.