Use FQ names in type mismatch error messages if necessary
In the previous implementation, the actual and expected types were
represented always by their simple names in type mismatch error
messages. As a result, if these simple names were equal, the shown error
message were difficult to understand (e.g. 'cannot convert from A to
A').

This commit changes this behavior such that fully qualified names are
used if the simple names of mismatching types are equal.

See #40.