Reschedule JtxtUML validation to the end of the build
When JtxtUML validation had been triggered by a Java build -- and not by
reconciliation --, in some cases created JtxtUML error markers was
removed by the cleaning process which occured during later phases of the
causing Java build.

This commit reschedules JtxtUML validation such that in case of a
build-related trigger, it is performed at the end of the current build
process. It should be noted that apparently, the used `buildStarting`
and `buildFinsihed` functions are called in pairs and for the same
project, therefore the introduced map-based caching is not necessary.
However, as I have found no guarantees that my observations regarding
the build behavior are correct, I have decided to choose the safer
approach.

See #426.