Extend case insensitive uniqueness checks to cross-file level
This solution is based on the reinterpretation of ResourceExceptions
which occur due to case-conflicting class files. As such, it is a bit
less responsive than standard validation, i.e. a build is required to
create error markers which are cleaned only during rebuilding (reconcile
is also sufficient, fortunately).

See d2f10e8 and #388.