Introduce a DSL for testing the XtxtUML parser
Unit tests of the XtxtUML parser are essentially AST comparisons: given
an XtxtUML source, we have to examine whether the AST produced by the
parser is equivalent to the expected one.

As there is no built-in support for this in Xtext, AST comparisons are
achievable only by fine-grained assert calls on the produced AST, where
the expected AST is embodied only indirectly by the expected values of
the individual assertions. Not only is this representation difficult to
interpret for a developer, maintainability is also an issue.

This commit introduces a slightly different, yet equivalent approach in
the form of a small, declarative DSL to provide a way of describing the
expected AST and the underlying assertions all at once, in the same
representation. Leveraging lambda functions and the expressive power of
Xtend, it is now possible to define unit tests for the XtxtUML parser in
a simple, intuitive and maintainable way.