Restore the default scope resolution process
Our customization of the scope resolution process was introduced in #89,
as an important part of making elements belonging to the same model
referable in action code by their partially qualified names. This
modification was necessary, because Xtext -- as it is described in #89
-- handles nested classes differently, which, combined with the fact
that back then txtUML models could be defined in a single enclosing
class, caused undesirable behavior.

With the introduction of split models in #163, this customization is no
longer needed, as there is no way to define a common enclosing class for
different nested model elements, which then would needed to be
referenced by each other in action code.

See 53c99b9.