Calculate expected error marker offsets dynamically
The previous implementation of validation unit tests defined expected
offsets of error markers as hard-coded integer literals. This made some
of the tests fail, as depending on the representation of newlines in the
given test environment, the actual error markers could appear on
slightly different absolute positions.

Note that the issue emerges from the fact that most of the test sources
for XtxtUML are defined as Xtend text literals, which are translated
on-the-fly using string builders.

This commit eliminates the aforementioned problem by calculating the
expected offsets at runtime by String.indexOf, which also makes the
cases more readable.