Run integration tests during non-release builds
Currently, the only difference between a non-release and a release build
is that the latter uses the 'release' Maven profile instead of the
default one. Until now, the 'release' profile specified that official
update sites shall be used instead of our mirrors for assembling the
target platform, and in addition to the default profile - according to
which only unit tests had to be performed -, integration tests are also
needed.

With the introduction of the `utils.eclipse.tests` project, non-release
Maven builds started to fail. The reason for this was that this project
only contained integration tests, which - according to the default
profile used during non-release builds - weren't run, and as a result,
the build failed, because the Maven Surefire plugin (used for running
tests during Maven builds) makes the build fail if no tests are present
in a test project.

As running integration tests during non-release builds adds only about
10% time overhead, we decided to include them in the default profile as
well.