Include MOXy in the generated update site
This is required because apparently, `discovery` tags in the feature.xml
don't really work. As a result, during a clean install Eclipse can't
resolve the external update site of MOXy and ultimately, the install
fails.

However, we can leverage that dependencies listed with match rule
`perfect` in the feature are automatically included in the generated
update site, therefore their availability is guaranteed.