The api.stdlib project is no longer auto-included on the classpath.
As the team agreed, the api.stdlib project will no longer serve as a plugin to be auto-included on the classpath of a txtUML project. Instead, it must be imported into the workspace and be added to the classpath of a project as a separate entry (if needed).

This commit adds a new option when creating a txtUML project whether to add this separate entry to the classpath or not.