Disable filtering of external type proposals
This commit tweaks the behavior of the content assist in case of
external elements. If the content assist is called where a type has to
be declared, either the default JDT-based or the custom txtUML-based
scope provider is used, depending on the externality of the enclosing
entity.

Please note that the issue discussed in #507 restricts the capabilities
of the content assist provider here as well. Additionally, due to
limitations of the grammar, in some cases the externality cannot be
detected even if the correct keyword is present. The reason for this is
that incomplete language constructs can't reliably be parsed, a problem
which could be solved with the method we used in 79a0255, but in this
case it is not possible, as making the causative `type` reference
optional in the declaration prefix rule would result in an ambiguous
grammar.