Add support for class attribute initializers in XtxtUML
With the addition of static attributes, the lack of support for
initializer expressions became quite inconvenient in XtxtUML. However,
as initializers are currently not supported by the model exporter, we
restrict their usage to external attributes only. This is still
sufficient, because for now, only external attributes can be static.