Remove api.stdlib from the feature
Due to #525 and #321, classes of the api.stdlib project have to be
defined in XtxtUML. Unfortunately, we can't automatically build and
package such classes with Maven yet, therefore the user has to import
the api.stdlib to their workspace separately for depending models.

The reason why we can't build XtxtUML files with Maven is the lack of a
dedicated XtxtUML Maven plugin (which would tell when and how to compile
these files), and as a result, the Maven build would fail if api.stdlib
was included in the process.

A previous commit, e2f6d3a, resolved the failing build by versioning the
files Eclipse generated from the XtxtUML classes of api.stdlib. However,
as this project will have to be imported to the workspace separately,
and also no other plugin depends on it, there's no need to include it in
the txtUML feature (and in the build process) for now.