Reflexive link arrangement
The GraphSearch class now has a new constructor which allows the specification of whether the actual link to be arranged is a reflexive link. If it is then instead of calculating a length cost for the link a new cost called 'reflexive link minimum length' is calculated which is the difference between the length of the link so far and the maximal distance between two possible endpoints.