Added new SignalWrapper type and renamed AssociationEndWrapper as AssociationEndRuntime. Also reworked the way someone is able to get information about the sender of a signal: Now the processingSignal trace method has a third, Optional<ModelClass> parameter which is the sender (if known).
The new SignalWrapper type wraps a signal, a sender ModelClass instance, and a sender port, all of which can be null. It simplifies the exhchanging of information in the execution project.

The rename was necessary to establish a proper naming convention in the modeling API and in the execution project: Wrappers which have a state (i.e., which are mutable) are called 'Runtime'.

The third parameter of the processingSignal method is necessary because in the old solution, the sender and the target were separated in the listeners and therefore they could not be unambiguously paired.