Fixed error with awaitTermination (see below); also changed the AbstractModelExecutor.OwnedThread class to ensure that all the threads are successfully registered before started.
The cause of the problem was that the performTermination() method was called in the unregisterThread() method outside the synchronization block.

Because of this the following scenario was possible:

	- The model executor starts.
	- The model executor thread registers (sync on threads list) and starts.
	- The thread finishes its work.
	- The thread unregisters (sync on threads list).
	- After releasing the lock on the threads list, the thread runs performTermination() and executes all the termination listeners.
	- Some time after the thread released the lock on the threads list, the main thread (which created and started the model executor) reaches the awaitTermination() call.
	- The main thread observes that the threads list is empty and happily returns.
	- At this point, even if the model executor thread has succeeded to execute the termination listeners, the main thread is unaware of the fact because no synchronization has happened between the two after the execution of the listeners.

Therefore, moving the performTermination call inside the synchronization block solves the problem. However, when searching for the problem, I modified the awaitTermination() method to use a CountDownLatch instead of joining the threads, and after I did find the cause of the problem, I kept this new solution because it seemed cleaner.