Initial implementation of the `connect` expression in XtxtUML
In order to keep related code as clean as possible, new abstractions
were introduced. Due to limitations of Xtext, special grammar rules had
to be constructed to achieve this. With the help of the added helper
class, which provides a common query interface for associations and
connectors, the existing implementation could mostly be reused.

This commit also addresses some minor fixes, namely in the compiler (see
the new adapter class -- it turned out tampering with the AST could
indeed be harmful) and in the import organizer.

See #385.