Fix Tycho build of the execution project
In 1916f03, a new dependency was introduced in the form of package
`com.sun.net.httpserver`. This can be found in most JRE implementations,
however, it's not part of the official API, and as such, is considered
unaccessible in Eclipse and during the Maven build.

It can be made accessible in the classpath configuration of the
referencing project, but, for some reason, Tycho doesn't detect this
during the build, and apparently, it's also not possible to manually
pass the custom access restriction as a compiler argument.

As a workaround, this commit makes Tycho ignore all access restriction
errors in the execution project, which method has been found as the
least undesirable one. Note that this could cause problems only if the
project is built without inspecting its code in Eclipse beforehand,
but this case is not to be expected.