Add element-name -> shape-id abstraction in JointJS
At the moment the animation backend serves the list of currently active
objects along their execution IDs and locations. The location is encoded
as the fully qualified name of the corresponding model element (e.g.
vertex).

In the frontend we use these qualified names to modify the appearance of
the active vertices, a method which works because JointJS embeds the
name as an attribute to the generated HTML. However, the ID attribute of
link shapes (e.g. transitions) is autogenerated for some reason, and we
couldn't find a way to change this.

This commit introduces an abstraction so that actual shape IDs -- which
can be used to select HTML elements that are active during the execution
-- can uniformly be obtained through the visualizer instance via their
fully qualified names.