Rename `Timer.start` to `Timer.schedule`
This is to make using timers in XtxtUML models more convenient, as
due to the newly added `start` language element, the previous
`Timer.start` method could be called only by escaping its name with a
`^` character. The phrase `schedule` could also be regarded more
expressive, because the purpose of the renamed method is to schedule a
send operation anyway.