Edit grammar rules for collections
The parser now correctly identifies collections declared inside of a
class, along with its modifiers and multiplicity. The current
implementation allows user defined class objects as well as basic
data types as the contained element. Note that currently, only part of
the grammar is present and collections are not yet fully supported.